Add some constants to FillOptions and StrokeOptions.
